refactor(midnight): extract shared connector build helpers into a build module
makeIntent and balanceSealedTransaction need the same recipient/token decoding, the
shared request DTOs (TransferKind, DesiredOutput), and the prove-to-unsealed-bytes
step that makeTransfer introduced. Pull them out of make_transfer.rs into a new
dapp_connector/build module so the sibling adapters reuse them instead of duplicating.

Pure move: makeTransfer keeps its own offer builders and parsing; behavior and tests
are unchanged.
